% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{agg_det_scores}
\alias{agg_det_scores}
\title{Aggregate deterministic scores}
\usage{
agg_det_scores(SCORENAME = NULL, RMSE = NULL, ME = NULL, MSE = NULL,
  SD = NULL, MAE = NULL, LEN = NULL)
}
\arguments{
\item{SCORENAME}{score name string}

\item{RMSE}{rmse scores of data subsets}

\item{ME}{me scores of data subsets}

\item{MSE}{mse scores of data subsets}

\item{SD}{sd scores of data subsets}

\item{MAE}{mae scores of data subsets}

\item{LEN}{length of forecast-observation pairs in subsets}
}
\value{
pooled score value
}
\description{
Aggregate deterministic scores
}
\examples{
x    = runif(1000)fnames   = system("ls ~/examplesRfdbk/icon/synop/*",intern=T)
y    = rnorm(1000)
x1   = x[1:10];x2=x[11:300];x3=x[301:1000]
y1   = y[1:10];y2=y[11:300];y3=y[301:1000]
rmse = function(x,y){return(sqrt(mean((x-y)^2)))}
rmse(x,y)
agg_det_scores("RMSE",RMSE=c(rmse(x1,y1),rmse(x2,y2),rmse(x3,y3)),LEN=c(length(x1),length(x2),length(x3)))
}
\author{
Felix <felix.fundel@dwd.de>
}
