% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{bitcheck}
\alias{bitcheck}
\title{Function evaluating the flags bit}
\usage{
bitcheck(int, bits)
}
\arguments{
\item{int}{a integer or a vector of integers that is converted to bits}

\item{bits}{a bit or a vector of bits that might be contained in each int}
}
\value{
logical if bit is in the converted integer
}
\description{
Some feedback file attributes are integers that have to be transformed to a bit string in order to encode the feedback file report (e,g, flags or level_sig).
This function takes the integers as input and a vector of bits and returns TRUE/FALSE if bit is part of the bit-converted integer
The first bit is 0 (not 1), in agreement with the construction of feedback files.
}
\examples{
bitcheck(1:100,0)
DT = data.table(flags=1:10)
DT[bitcheck(flags,0)]
}
\author{
Felix <felix.fundel@dwd.de>
}
