% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{fdbk_dt}
\alias{fdbk_dt}
\title{Fdbk file content (as obtained from read_fdbk(_f)) is converted into a data.table.
Therefore a lot of data overhead is created as most data will be duplicated.
However, data.tables offer a lot of extra functionality.}
\usage{
fdbk_dt(fdbk)
}
\arguments{
\item{fdbk}{output from read_fdbk}
}
\value{
a data.table of the feedback file data section
}
\description{
Fdbk file content (as obtained from read_fdbk(_f)) is converted into a data.table.
Therefore a lot of data overhead is created as most data will be duplicated.
However, data.tables offer a lot of extra functionality.
}
\examples{
fdbk = read_fdbk("~/examplesRfdbk/icon/synop/verSYNOP.2014120112")
format(object.size(fdbk),"Mb")
DT   = fdbk_dt(fdbk)
format(object.size(DT),"Mb")
DT
}
\author{
Felix <felix.fundel@dwd.de>
}
