% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{fdbk_dt_brier}
\alias{fdbk_dt_brier}
\title{Calculate the brier score (and decomposition and skill score) for one threshold per variable}
\usage{
fdbk_dt_brier(DT, thresholds = "", by = "")
}
\arguments{
\item{DT}{data.table (columns 'veri_ens_member','obs' and 'veri_data' plus all variables to make forecasts distinguishable are required!!!)}

\item{thresholds}{list of threshold for variable names in DT (if "" uses obs median)}

\item{by}{stratify crps by (e.g. c('varno','veri_forecast_time'))}
}
\value{
data.table with columns as defined in 'by' plus scorename plus score
}
\description{
Calculate the brier score (and decomposition and skill score) for one threshold per variable
}
\examples{
require(ggplot2)
fnames    = system("/bin/ls ~/examplesRfdbk/eps/*12/verSYNOP*",intern=T)
condition = list(veri_description="grepl('member',veri_description)",
                 state="state\%in\%c(0,1)",
                 statid="!is.na(as.numeric(statid)) & !duplicated(statid)",
                 veri_forecast_time="veri_forecast_time>=1200")
columns   = c("obs","veri_data","varno","veri_ens_member","veri_forecast_time","statid","veri_initial_date")
DT        = fdbk_dt_multi_large(fnames,condition,columns,5)
PROBS     = fdbk_dt_brier(DT,by=c("varno","veri_forecast_time"))

ggplot(PROBS,aes(x=veri_forecast_time,y=score,color=scorename,group=scorename))+
  geom_line()+geom_point()+facet_wrap(~varno,scale="free_y",ncol=2)+
  theme_bw()+scale_colour_discrete("lead-time")
}
\author{
Felix <felix.fundel@dwd.de>
}
