% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{fdbk_dt_conttable}
\alias{fdbk_dt_conttable}
\title{Calculates stratified contingency table entries (above threshold) for a data table}
\usage{
fdbk_dt_conttable(DT, vars = NULL, thrs = NULL, by = NULL, cores = 1)
}
\arguments{
\item{DT}{data.table with relevant information}

\item{vars}{character vector of varnos (if NULL take from DT)}

\item{thrs}{list of vectors of thresholds for each varno (if NULL threshold are generated from quantiles)}

\item{by}{stratify contingency entries by these DT columns}

\item{cores}{number of CPU cores to split the calculation (helps for larger data tables)}
}
\value{
data.table with columns varno,thr, hits,false,miss,corrneg and the arguments of 'by'
}
\description{
Calculates stratified contingency table entries (above threshold) for a data table
}
\examples{
#EXAMPLE (CSI for quantile thresholds) 
require(ggplot2)
fnames                = system("ls ~/examplesRfdbk/*/synop/verSYNOP.*",intern=T)
cond                  = list(veri_description="grepl('forecast',veri_description)",
                             veri_forecast_time="veri_forecast_time\%in\%c(2400,4800,7200,9600,12000,14400,16800)")
columnnames           = c("obs","veri_data","varno","veri_model","veri_forecast_time","statid","veri_initial_date")
DT                    = fdbk_dt_multi_large(fnames,cond,columnnames,20)
vars                  = c('1','3','4','29')
thrs                  = list('1'=c(50,60),'3'=c(-5,0,5),'4'=c(-5,0,5),'29'=c(.4,.6,.8))
xx                    = fdbk_dt_conttable(DT,vars=vars,thrs=thrs,by=c("veri_model","veri_forecast_time","varno"),cores=10)
CSI                   = xx[,list(csi =(hit)/(hit + miss + false) ),by=c("veri_forecast_time","veri_model","thr","varno")]
CSI[,varno:=varno_to_name(varno,T)]
ggplot(CSI,aes(x=thr,y=csi,color=factor(veri_forecast_time),linetype=factor(veri_model),group=interaction(varno,veri_forecast_time,veri_model)))+
geom_line()+
ggtitle("CSI")+
facet_wrap(~varno,scales="free_x")
}
\author{
Felix <felix.fundel@dwd.de>
}
