% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{fdbk_dt_conttable_2thrs}
\alias{fdbk_dt_conttable_2thrs}
\title{Calculates stratified contingency table entries (above or between thresholds) for a data table}
\usage{
fdbk_dt_conttable_2thrs(DT, thrs, by, cores = 1, incores = 1)
}
\arguments{
\item{DT}{data.table with relevant information (at least varno, obs and veri_data)}

\item{thrs}{list of variable having each a list of lower/upper thresholds (set upper to Inf if only one threshold is required)}

\item{by}{stratify contingency entries by these DT columns}

\item{cores}{computing cores for the outer loop (splits computation by varnos)}

\item{incores}{computing cores for the outer loop (splits computation by thresholds)(available cores have to be of number cores x incores)}
}
\value{
data.table with columns varno,thr, hits,false,miss,corrneg and the arguments of 'by'
}
\description{
Calculates stratified contingency table entries (above or between thresholds) for a data table
}
\examples{
#EXAMPLE (CSI for quantile thresholds) 
require(ggplot2)
fnames                = system("ls ~/examplesRfdbk/*/synop/verSYNOP.*",intern=T)
cond                  = list(veri_description="grepl('forecast',veri_description)",
                             veri_forecast_time="veri_forecast_time\%in\%c(2400,4800,7200,9600,12000,14400,16800)")
columnnames           = c("obs","veri_data","varno","veri_model","veri_forecast_time","statid","veri_initial_date")
DT                    = fdbk_dt_multi_large(fnames,cond,columnnames,20)
thrs                  = list('29'=list('lower'=c(.5,.8),'upper'=c(Inf,.9)),
                              '3'=list('lower'=c(-5,0,5),'upper'=c(Inf,Inf,Inf)))
xx                    = fdbk_dt_conttable_2thrs(DT,thrs,by=c("veri_model","veri_forecast_time","varno"),cores=2,incores=3)
CSI                   = xx[,list(csi =(hit)/(hit + miss + false) ),by=c("veri_forecast_time","veri_model","thr","varno")]
CSI[,varno:=varno_to_name(varno,T)]
ggplot(CSI,aes(x=veri_forecast_time,y=csi,group=interaction(veri_model,thr),linetype=veri_model,color=thr))+
geom_line()+
facet_grid(~varno)+
ggtitle("CSI")
}
\author{
Felix <felix.fundel@dwd.de>
}
