% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fdbk_dt_multi}
\alias{fdbk_dt_multi}
\title{Load relevant information of many feedback files as data.table
Be restrictive with the columns kept in the data.table as otherwise
the memory limit is  reached fast
To speed up computation multiple cores are utilized (if possible)}
\usage{
fdbk_dt_multi(fnames, cond = "", columnnames = "", cores = 1)
}
\arguments{
\item{fnames}{vector of feedback filename(s)}

\item{cond}{string of conditions the fdbk file will be filtered for in advance}

\item{columnnames}{attribute names to keep in the data table}
}
\value{
a data.table of merged feedback file contents
}
\description{
Load relevant information of many feedback files as data.table
Be restrictive with the columns kept in the data.table as otherwise
the memory limit is  reached fast
To speed up computation multiple cores are utilized (if possible)
}
\examples{
fnames      = system("ls ~/examplesRfdbk/icon/synop/verSYNOP.*",intern=T)
cond        = "varno\%in\%c(3,4) & !is.na(obs)"
columnnames = c("obs","veri_data","varno","veri_forecast_time")
DT          = fdbk_dt_multi(fnames,cond,columnnames,4)
DT
}
\author{
Felix <felix.fundel@dwd.de>
}

