% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fdbk_dt_verif_continuous_windDir}
\alias{fdbk_dt_verif_continuous_windDir}
\title{Deterministic scores for wind direction in degrees with bootstrap confidence intervals if required}
\usage{
fdbk_dt_verif_continuous_windDir(DT, strat, bootscores = F, R = 100)
}
\arguments{
\item{DT}{data table (obs and veri_data are required,only for wind direction in degrees!)}

\item{strat}{list of variables to stratify for}

\item{bootscores}{logical if bootstrap confidence intervals are required (5-95)}

\item{R}{number of bootstrap iterations (default 100)}
}
\value{
a data.table of stratified continuous verification scores  (ME,SD,RMSE,R2,LEN)
}
\description{
Deterministic scores for wind direction in degrees with bootstrap confidence intervals if required
}
\author{
Felix <felix.fundel@dwd.de>
}

