% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{fdbk_dt_wide}
\alias{fdbk_dt_wide}
\title{Function returning ensemble feedback file content as data.table with member forecasts as individual columns}
\usage{
fdbk_dt_wide(fdbk)
}
\arguments{
\item{fdbk}{return from read_fdbk or read_fdbk_larg}
}
\value{
a data.table
}
\description{
Having the forecasts from each member as column (wide table) instead of a single veri_data column (long table) has
a great benefit for large ensembles and many observations. Wide tables carry less redundant data, use less memory
and faster computation is possible
}
\examples{
fdbk_dt_wide(read_fdbk_large(filename,cond,vars))
}
\author{
Felix <felix.fundel@dwd.de>
}
