% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{fdbk_refdate}
\alias{fdbk_refdate}
\title{Get reference date(s) from feedback file(s)}
\usage{
fdbk_refdate(filenames)
}
\arguments{
\item{filenames}{filename(s) fo feedback file(s) inluding path}
}
\value{
vector of reference dates YYYYmmddHHMM
}
\description{
Get reference date(s) from feedback file(s)
}
\examples{
filenames  = system("ls ~/examplesRfdbk/icon/synop/*",intern=T)
fdbk_refdate(filenames)
}
\author{
Felix <felix.fundel@dwd.de>
}
