% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{hhmm2hour}
\alias{hhmm2hour}
\title{Function to convert time in format hhmm to decimal hours.
Useful to calculate a derived time from two time informations}
\usage{
hhmm2hour(x)
}
\arguments{
\item{time}{in format hhmm. Can be a string or numeric}
}
\value{
time in decimal hours
}
\description{
Function to convert time in format hhmm to decimal hours.
Useful to calculate a derived time from two time informations
}
\examples{
#EXAMPLE 1 simple examples
hhmm2hour(0145) returns 1.75
hhmm2hour(145) returns 1.75
hhmm2hour("145") returns 1.75

#EXAMPLE 2 calculate leadtime from veri_forecast_time and time
require(ggplot2)
fnames       = "/Users/josuegehring/Desktop/verTEMP.2014120112"
cond        = list(obs="!is.na(obs)",varno="varno\%in\%c(2,3,4,29)",ident="ident\%in\%c(6610)",varno="varno\%in\%c(2)")
columnnames = c("obs","veri_data","varno","state","level","veri_forecast_time","time","ident")
DT          = fdbk_dt_multi_large(fnames,cond,columnnames,1)
leadtime = hhmm2hour(DT$veri_forecast_time) + DT$time/60
DT[,"leadtime"] = leadtime
}
\author{
Felix <felix.fundel@dwd.de>
}

