% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{lonlat_to_synopregion}
\alias{lonlat_to_synopregion}
\title{Non-overlapping regions, specifically defined for the DWD SYNOP verification}
\usage{
lonlat_to_synopregion(lon, lat)
}
\arguments{
\item{lon}{longitude vector}

\item{lat}{latitude vector}
}
\value{
a vector of same lenght as lon or lat with character strings of the region for each point, NA for no match
}
\description{
Non-overlapping regions, specifically defined for the DWD SYNOP verification
}
\examples{
DT = data.table(lon=c(15,85),lat=c(-30,40))
DT[,region:=lonlat_to_synopregion(lon,lat)]
DT
}
\author{
Felix <felix.fundel@dwd.de>
}
