% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_fdbk}
\alias{read_fdbk}
\title{Load the entire content of a fdbk file}
\usage{
read_fdbk(filename)
}
\arguments{
\item{filename}{NetCDF fdbk filename including path}
}
\value{
a list of entries from the given fdbk file
}
\description{
Load the entire content of a fdbk file
}
\examples{
fdbk = read_fdbk("~/examplesRfdbk/icon/synop/verSYNOP.2014120112")
str(fdbk)
}
\author{
Felix <felix.fundel@dwd.de>
}

