% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{read_fdbk_large}
\alias{read_fdbk_large}
\title{Load one fdbk file and return as list of lists of....
condition and vars arguments help to discard data you do not need}
\usage{
read_fdbk_large(fname, condition = "", vars = "")
}
\arguments{
\item{fname}{feedback filename (including path)}

\item{condition}{list of strings of conditions (all of the list entries are connected with the "&" operator!)}

\item{vars}{vector of variable names that should be retained if not specified or "" all variables are loaded}
}
\value{
a data.table with fdbk file content
}
\description{
Load one fdbk file and return as list of lists of....
condition and vars arguments help to discard data you do not need
}
\examples{
#EXAMPLE 1 (1x1 deg.) bias of satellite data (channel 921 from METOP-1)
fnames      = "~/examplesRfdbk/example_monRad/monRAD_2014092406.nc"
condition   = list(obs="!is.na(obs)",
                   level="level\%in\%c(921)",
                   statid="statid=='METOP-1   '",
                   veri_forecast_time="veri_forecast_time==0",
                   veri_run_type="veri_run_type==3",
                   veri_ens_member="veri_ens_member==-1")
fdbk         = read_fdbk_large(fnames,condition,c("lon","lat","obs"))
x11(width=12,height=7.5)
scatterplot(fdbk$DATA$lon$values,fdbk$DATA$lat$values,fdbk$DATA$obs$values,pch=20,cex=.5,cpal=tim.colors(),ncol=30);world(add=T)
}
\author{
Felix <felix.fundel@dwd.de>
}
