% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{scatterplot}
\alias{scatterplot}
\title{Scatterplot with colored points}
\usage{
scatterplot(x, y, z, zlim = NULL, ncol = 10, cpal = c("red", "white",
  "blue"), ...)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{z}{numeric vector}

\item{zlim}{plot color range (default z range)}

\item{ncol}{number of colors (default 10)}

\item{cpal}{color palette (default red,white,blue)}
}
\value{
a plot
}
\description{
Scatterplot with colored points
}
\examples{
condition  = list(obs="!is.na(obs)",level="level\%in\%c(921)",statid="statid=='METOP-1   '",veri_forecast_time="veri_forecast_time==0",veri_run_type="veri_run_type==3",veri_ens_member="veri_ens_member==-1")
DT         = fdbk_dt_multi_large("~/examplesRfdbk/example_monRad/monRAD_2014092406.nc",condition,vars=c("obs","lon","lat"),cores=1)
x11(width=12,height=7.5)
DT[,scatterplot(lon,lat,obs,pch=20,cpal=tim.colors(),ncol=20,cex=.5)]
world(add=T,col="gray",fill=T)
}
\author{
Felix <felix.fundel@dwd.de>
}
