% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{statid_to_wmoregion}
\alias{statid_to_wmoregion}
\title{Convert WMO station-id to region}
\usage{
statid_to_wmoregion(ident)
}
\arguments{
\item{ident}{numeric vector of station ID as integer (see variable "ident" in feedback file)}
}
\value{
vector of same length wiith id replaced by region shortcut
}
\description{
Convert WMO station-id to region
}
\author{
Felix <felix.fundel@dwd.de>
}
