% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{varno_to_name}
\alias{varno_to_name}
\title{Convert variable number (varno) to long or short variable name and reverse}
\usage{
varno_to_name(varno, short = T, rev = F)
}
\arguments{
\item{short}{short or long name (boolean)}

\item{rev}{TRUE: from varno to name, FALSE: from short name to varno}

\item{varno(s)}{or short name(s)}
}
\value{
long or short variable name(s)
}
\description{
Convert variable number (varno) to long or short variable name and reverse
}
\examples{

varno_to_name(c(3,4),short=T,rev=F)
varno_to_name(c(3,4),short=F,rev=F)
varno_to_name(c("RH","TS"),short=T,rev=T)
varno_to_name(c("RH","TS"),short=F,rev=T)
varno_to_name("geopotential (m^2/s^2)",short=F,rev=T)
varno_to_name(varno_to_name("geopotential (m^2/s^2)",short=F,rev=T))
}
\author{
Felix <felix.fundel@dwd.de>
}
