% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdbk_asdataframe.R
\name{windBias}
\alias{windBias}
\title{Difference in wind direction (based un U. Pflügers code)}
\usage{
windBias(ang_pred, ang_obs)
}
\arguments{
\item{ang_pred}{forecast wind direction}

\item{ang_obs}{observed wind direction}
}
\value{
wind direction difference in degree
}
\description{
Difference in wind direction (based un U. Pflügers code)
}
\author{
Felix <felix.fundel@dwd.de>
}
