% Generated by roxygen2 (4.0.2): do not edit by hand
\name{windDir}
\alias{windDir}
\title{Convert u,v wind in wind direction in degrees}
\usage{
windDir(u, v)
}
\arguments{
\item{u}{u wind vector}

\item{v}{v wind vector}
}
\value{
wind direction in degree (0 - <360), 360 is set to 0, if u&v=0 then return NA
}
\description{
Convert u,v wind in wind direction in degrees
}
\examples{
u = c( 10,   0,   0, -10,  10,  10, -10, -10, 0)
v = c(  0,  10, -10,   0,  10, -10,  10, -10, 0)
windDir(u,v)
}
\author{
Felix <felix.fundel@dwd.de>
}

