% Generated by roxygen2 (4.0.2): do not edit by hand
\name{windSpeed}
\alias{windSpeed}
\title{Convert u,v wind in wind speed}
\usage{
windSpeed(u, v)
}
\arguments{
\item{u}{u wind vector}

\item{v}{v wind vector}
}
\value{
wind speed
}
\description{
Convert u,v wind in wind speed
}
\author{
Felix <felix.fundel@dwd.de>
}

