% Generated by roxygen2 (4.0.2): do not edit by hand
\name{afc}
\alias{afc}
\title{Fast version of the 2AFC for continuous observations and forecasts
The score is based on the rank correlation coefficient}
\usage{
afc(obsv, fcst)
}
\arguments{
\item{obsv}{observation vector}

\item{fcst}{forecast vector}
}
\value{
afc score
}
\description{
Fast version of the 2AFC for continuous observations and forecasts
The score is based on the rank correlation coefficient
}
\author{
Felix <felix.fundel@dwd.de>
}

