% Generated by roxygen2 (4.0.2): do not edit by hand
\name{windBias}
\alias{windBias}
\title{Difference in wind direction (based un U. Pflügers code)}
\usage{
windBias(ang_pred, ang_obs)
}
\arguments{
\item{ang_pred}{forecast wind direction}

\item{ang_obs}{observed wind direction}
}
\value{
wind direction difference in degree
}
\description{
Difference in wind direction (based un U. Pflügers code)
}
\author{
Felix <felix.fundel@dwd.de>
}

